+++ VGraph.csproj
@@ -138,6 +138,7 @@
     <Compile Include="src\drawTools\LineTool.cs" />
     <Compile Include="src\objects\History.cs" />
     <Compile Include="src\objects\LineSegment.cs" />
+    <Compile Include="src\ui\MainWindowCommands.cs" />
     <Compile Include="src\ui\MenuBarControl.xaml.cs">
       <DependentUpon>MenuBarControl.xaml</DependentUpon>
     </Compile>
